# Optimize unique constraints operation time checks for write queries