--- a/concept/thing/attribute.rs
+++ b/concept/thing/attribute.rs
@@ -10,6 +10,7 @@ use std::{
     fmt,
     hash::{Hash, Hasher},
     iter,
+    ops::RangeBounds,
     sync::{Arc, OnceLock},
 };
 
@@ -39,8 +40,12 @@ use storage::{
 use crate::{
     edge_iterator,
     error::{ConceptReadError, ConceptWriteError},
-    thing::{object::Object, thing_manager::ThingManager, HKInstance, ThingAPI},
-    type_::{attribute_type::AttributeType, ObjectTypeAPI},
+    thing::{
+        object::{HasReverseIterator, Object},
+        thing_manager::ThingManager,
+        HKInstance, ThingAPI,
+    },
+    type_::{attribute_type::AttributeType, object_type::ObjectType, ObjectTypeAPI},
     ConceptAPI, ConceptStatus,
 };
 
@@ -91,6 +96,15 @@ impl Attribute {
         owner_type: impl ObjectTypeAPI,
     ) -> AttributeOwnerIterator {
         thing_manager.get_owners_by_type(snapshot, self, owner_type)
+    }
+
+    pub fn get_has_by_owner_types(
+        &self,
+        snapshot: &impl ReadableSnapshot,
+        thing_manager: &ThingManager,
+        owner_type_range: &impl RangeBounds<ObjectType>,
+    ) -> HasReverseIterator {
+        thing_manager.get_has_reverse_by_attribute_and_owner_type_range(snapshot, self, owner_type_range)
     }
 
     pub fn next_possible(&self) -> Attribute {