## Release notes: product changes
Modify the algorithm of attribute uniqueness validation at operation time, performed for each inserted `has` instance of ownerships with `@unique` and `@key` annotations. 
These changes boost the performance of insert operations for constrained `has`es by up to 35 times, resulting in an almost indiscernible effect compared to the operations without uniqueness checks. 

## Motivation
Uniqueness validations for inserts had a disastrous effect on inserts' performance.

## Implementation
The old algorithm iterated over all objects of the suitable object types (with `owns` under the constraint verified). Then, each object was verified as not having an attribute with the value verified. This straightforward search took more and more time with the increased number of objects in the database, even if they didn't have common attributes: it was bound only to the schema.

The new algorithm performs the reversed sequence of actions. First, it searches for attributes with the value verified. Then, these attributes (they can be of different attribute types) are checked on the existence of `has` edges to objects of the verified set of types. This way, we limit the number of elements and storage lookups processed, making better use of the data specifics and the iterator optimizations.

Using a shortened version of [a small benchmark](https://discord.com/channels/665254494820368395/672580553677078548/1354875257382637830), which performs simple insert operations for objects with randomized attributes, we found that a single average insert operation took (*only use this data for relative comparison: these results were gathered in a local environment*):
Non-optimal compilation mode (x35 times better than before):
* 35.65 ms in the old version with uniqueness validations 
* 1.07 ms without uniqueness validations
* 1.09 ms in the new version with uniqueness validations

Optimal compilation mode (x25 times better than before):
* 3.77 ms in the old version with uniqueness validations
* 0.15 ms without uniqueness validations
* 0.15 ms in the new version with uniqueness validations